Other changes associated with motion planner

+++ rationality/util.py
@@ -24,20 +24,21 @@ def ordinal(n: int) -> str:
     return "%d%s" % (n, 'tsnrhtdd' [(n // 10 % 10 != 1) * (n % 10 < 4) * n % 10::4])
 
 
-def savefig(base_name: str, formats: Iterable[str], pad=0.0) -> None:
-    for format in formats:
-        filename = f'{base_name}.{format}'
-        print(f'Saving {format.upper()} format: {filename}')
+def savefig(base_name: str, formats: Iterable[str], pad=0.0, perform_save=True) -> None:
+    if perform_save:
+        for format in formats:
+            filename = f'{base_name}.{format}'
+            print(f'Saving {format.upper()} format: {filename}')
 
-        if os.path.exists(filename):
-            os.remove(filename)
+            if os.path.exists(filename):
+                os.remove(filename)
 
-        if format in {'pgf'}:
-            plt.savefig(f'{base_name}.{format}', format=format, bbox_inches='tight', pad_inches=pad, backend='pgf')
-        elif format in {'eps', 'jpeg', 'jpg', 'pdf', 'pgf', 'png', 'ps', 'raw', 'rgba', 'svg', 'svgz', 'tif', 'tiff'}:
-            plt.savefig(f'{base_name}.{format}', format=format, bbox_inches='tight', pad_inches=pad)
-        else:
-            raise ValueError(f"Unknown file format: `{format}'.")
+            if format in {'pgf'}:
+                plt.savefig(f'{base_name}.{format}', format=format, bbox_inches='tight', pad_inches=pad, backend='pgf')
+            elif format in {'eps', 'jpeg', 'jpg', 'pdf', 'pgf', 'png', 'ps', 'raw', 'rgba', 'svg', 'svgz', 'tif', 'tiff'}:
+                plt.savefig(f'{base_name}.{format}', format=format, bbox_inches='tight', pad_inches=pad)
+            else:
+                raise ValueError(f"Unknown file format: `{format}'.")
 
 
 @jax.jit
